refactor: 채팅 인포 탭 코드 분리
