+++ src/pages/chat/components/sidebar/channel-property-content.tsx
