+++ src/pages/chat/components/sidebar/chat-sidebar-for-admin.tsx
@@ -1,5 +1,5 @@
 import { Tabs } from 'flowbite-react';
-import ChatInformationTap from 'pages/chat/components/sidebar/chat-information-tap';
+import ChatInformationTap from 'pages/chat/components/sidebar/information-tap/chat-information-tap';
 import ChatMemberTap from 'pages/chat/components/sidebar/chat-member-tap';
 import ChatSettingTap from 'pages/chat/components/sidebar/setting-tap/chat-setting-tap';
 import { FiInfo, FiSettings, FiUsers } from 'react-icons/fi';
